:bug: fix returning elements on authenticated users

+++ Server/FlowHood/src/main/java/org/galactic/flowhood/controller/RequestController.java
@@ -1,18 +1,27 @@
 package org.galactic.flowhood.controller;
 
+import jakarta.validation.Valid;
 import org.galactic.flowhood.domain.dto.request.HouseReqDTO;
+import org.galactic.flowhood.domain.dto.request.RequestReqDTO;
 import org.galactic.flowhood.domain.dto.request.RequestStateReqDTO;
 import org.galactic.flowhood.domain.dto.response.GeneralResponse;
 import org.galactic.flowhood.domain.entities.Request;
+import org.galactic.flowhood.domain.entities.Role;
+import org.galactic.flowhood.domain.entities.User;
 import org.galactic.flowhood.services.RequestService;
+import org.galactic.flowhood.services.RoleService;
+import org.galactic.flowhood.services.UserService;
+import org.galactic.flowhood.utils.SystemRoles;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 import java.util.UUID;
 
+//TODO
 @RestController
 @RequestMapping("/api/request")
 @CrossOrigin
@@ -20,8 +29,14 @@ public class RequestController {
     final
     RequestService requestService;
 
-    public RequestController(RequestService requestService) {
+    final UserService userService;
+
+    final RoleService roleService;
+
+    public RequestController(RequestService requestService, UserService userService, RoleService roleService) {
         this.requestService = requestService;
+        this.userService = userService;
+        this.roleService = roleService;
     }
 
     //admin only
@@ -54,11 +69,13 @@ public class RequestController {
         }
     }
 
-    //TODO: think here
+    //Access on resident and responsible
     @PostMapping("/")
-    public ResponseEntity<GeneralResponse> createRequest(@RequestBody HouseReqDTO req){
+    public ResponseEntity<GeneralResponse> createRequest(@RequestBody @Valid RequestReqDTO req, BindingResult errors){
         //validate if role is vigilant or responsible then request must be automatically approved else it should be pending
         try{
+            if (errors.hasErrors())
+                return GeneralResponse.builder().status(HttpStatus.BAD_REQUEST).message("invalid request").data(errors.getAllErrors()).getResponse();
             return GeneralResponse.builder().status(HttpStatus.OK).message("role created").getResponse();
 
         }
@@ -67,6 +84,7 @@ public class RequestController {
         }
     }
 
+    //check if request is from resident or admin
     @DeleteMapping("/{_id}")
     public ResponseEntity<GeneralResponse> deleteRequest(@PathVariable("_id") String id){
         try{
@@ -74,14 +92,28 @@ public class RequestController {
             Request request = requestService.findRequestById(_id);
             if (request == null)
                 return GeneralResponse.builder().status(HttpStatus.NOT_FOUND).message("not found").getResponse();
-            requestService.deleteRequest(request);
-            return GeneralResponse.builder().status(HttpStatus.OK).message("deleted!").getResponse();
+
+            User user = userService.findUserAuthenticated().toEntity();
+
+            Role admRole = roleService.findRoleById(SystemRoles.ADMINISTRATOR.getRole());
+            Role resRole = roleService.findRoleById(SystemRoles.RESPONSIBLE.getRole());
+
+            if(user.getRoles().contains(admRole) ||
+                    (user.getRoles().contains(resRole) && request.getResident()
+                            .getId().equals(user.getId()))){
+                requestService.deleteRequest(request);
+                return GeneralResponse.builder().status(HttpStatus.OK).message("deleted!").getResponse();
+            }
+
+            return GeneralResponse.builder().status(HttpStatus.UNAUTHORIZED).message("could not be deleted!").getResponse();
 
         }
         catch (Exception e){
             return GeneralResponse.builder().status(HttpStatus.INTERNAL_SERVER_ERROR).getResponse();
         }
     }
+
+    //TODO
     @PatchMapping("/{_id}") //getting user validation and status of current request
     public ResponseEntity<GeneralResponse> changeStatus(@PathVariable("_id") String id, @RequestBody RequestStateReqDTO req){
         try{
